refactor: move contract file outside src folder
