+++ client/tsconfig.json
@@ -2,10 +2,17 @@
   "compilerOptions": {
     "target": "ES2020",
     "useDefineForClassFields": true,
-    "lib": ["ES2020", "DOM", "DOM.Iterable", "ES2024.Object"],
+    "lib": [
+      "ES2020",
+      "DOM",
+      "DOM.Iterable",
+      "ES2024.Object"
+    ],
     "module": "ESNext",
     "skipLibCheck": true,
-    "types": ["node"],
+    "types": [
+      "node"
+    ],
     /* Bundler mode */
     "outDir": "${configDir}/tsbuild",
     "moduleResolution": "bundler",
@@ -22,9 +29,17 @@
     "rootDir": "../",
     "baseUrl": "./",
     "paths": {
-      "@/*": ["./src/*"],
-      "@core/*": ["../core/*"]
+      "@/*": [
+        "./src/*"
+      ],
+      "@core/*": [
+        "../core/*"
+      ]
     }
   },
-  "include": ["./src/**/*", "./vite.config.ts"]
+  "include": [
+    "./src/**/*",
+    "./vite.config.ts",
+    "./contract.ts"
+  ]
 }
